# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PocketSync API
  description: |
    Consolidated personal finance manager API for the Nigerian market.

    **Production base URL:** `https://pocketsync.onrender.com`

    ## Authentication
    PocketSync uses **HttpOnly cookies** — not Bearer tokens. After `POST /api/v1/auth/login`,
    the server sets `accessToken` and `refreshToken` cookies. All protected routes require
    the browser/client to send cookies (`credentials: 'include'` in fetch, `withCredentials: true` in axios).

    ## Sign-up & password reset (email OTP)
    1. `POST /api/v1/auth/register` — creates account, sends 6-digit email OTP
    2. `POST /api/v1/auth/verify-otp` with `purpose: signup` — marks email verified
    3. `POST /api/v1/auth/login` — blocked until email is verified (`403`)
    4. Forgot password: `POST /api/v1/auth/forgot-password` → `POST /api/v1/auth/reset-password`

    OTP resend cooldown is **30 seconds**. Codes expire in **10 minutes**.
    Without `RESEND_API_KEY`, OTPs are logged to the server console; in development the code
    may also appear in the JSON response as `devOtp`.

    ## BVN onboarding (mocked — Phase 2)
    After email verification and login, complete onboarding:
    1. `GET /api/v1/onboarding/status` — route user to the correct step
    2. `POST /api/v1/onboarding/bvn/submit` — mock BVN + phone validation, sends SMS OTP
    3. `POST /api/v1/onboarding/bvn/verify-otp` — verifies phone, discovers mock accounts
...

...

    ## Amounts
...

...
servers:
...
paths:
...
  /api/v1/auth/send-otp:
    post:
      tags: [Auth]
      summary: Resend email verification code
      description: |
        Resends a 6-digit OTP for `signup` or `reset` purposes.
        Subject to a 30-second resend cooldown per email + purpose.
      operationId: sendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpRequest'
            example:
              email: user@example.com
              purpose: signup
      responses:
        '200':
          description: Verification code sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpSentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Account not found (signup purpose only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Resend cooldown active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpCooldownError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify email OTP
      description: |
        Validates a 6-digit code. For `purpose: signup`, marks the user's email as verified.
        For `purpose: reset`, only validates the code (use `reset-password` to set a new password).
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
            example:
              email: user@example.com
              code: '123456'
              purpose: signup
      responses:
        '200':
          description: Code accepted
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SignupVerifyResponse'
                  - $ref: '#/components/schemas/ResetVerifyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset code
      description: |
        Always returns the same success message whether or not the email exists (prevents enumeration).
        Sends a 6-digit OTP with `purpose: reset` when the account exists.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
            example:
              email: user@example.com
      responses:
        '200':
          description: Generic success message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with OTP
      description: Verifies the reset OTP and sets a new password. Invalidates existing refresh tokens.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
            example:
              email: user@example.com
              code: '123456'
              newPassword: NewSecurePass123
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password reset successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HttpOnly cookie set by POST /api/v1/auth/login
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          description: Present in development only
...
    SendOtpRequest:
      type: object
      required: [email, purpose]
      properties:
        email:
          type: string
          format: email
        purpose:
          type: string
          enum: [signup, reset]
    VerifyOtpRequest:
      type: object
      required: [email, code, purpose]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: '^\d{6}$'
        purpose:
          type: string
          enum: [signup, reset]
    OtpSentResponse:
      type: object
      properties:
        message:
          type: string
        resendAvailableIn:
          type: integer
        devOtp:
          type: string
          description: Present in development when email is mocked
    OtpCooldownError:
      type: object
      properties:
        error:
          type: string
        resendAvailableIn:
          type: integer
    SignupVerifyResponse:
      type: object
      properties:
        message:
          type: string
        emailVerified:
          type: boolean
          example: true
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            fullName:
              type: string
            bvnVerified:
              type: boolean
    ResetVerifyResponse:
      type: object
      properties:
        message:
          type: string
        verified:
          type: boolean
          example: true
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ForgotPasswordResponse:
      type: object
      properties:
        message:
          type: string
          example: If an account exists for this email, a verification code has been sent
        devOtp:
          type: string
          description: Present in development when account exists and email is mocked
    ResetPasswordRequest:
      type: object
      required: [email, code, newPassword]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: '^\d{6}$'
        newPassword:
          type: string
          minLength: 8
...